+++ example/linear_algebra.cpp
@@ -30,10 +30,10 @@
 namespace {
 
 template<class ET, class OT>
-std::ostream& operator<<(std::ostream& os, const std::experimental::math::vector<ET, OT>& v)
+std::ostream& operator<<(std::ostream& os, const std::math::vector<ET, OT>& v)
 {
   os << "|";
-  for (size_t i = 0;  i < v.size();  ++i) {
+  for (auto i = 0;  i < v.size();  ++i) {
     os << fmt::format(" {:>9}", v(i));
   }
   os << " |";
@@ -41,11 +41,11 @@ std::ostream& operator<<(std::ostream& os, const std::experimental::math::vector
 }
 
 template<class ET, class OT>
-std::ostream& operator<<(std::ostream& os, const std::experimental::math::matrix<ET, OT>& v)
+std::ostream& operator<<(std::ostream& os, const std::math::matrix<ET, OT>& v)
 {
-  for (size_t i = 0;  i < v.rows();  ++i) {
+  for (auto i = 0;  i < v.rows();  ++i) {
     os << "|";
-    for (size_t j = 0;  j < v.columns();  ++j) {
+    for (auto j = 0;  j < v.columns();  ++j) {
       os << fmt::format(" {:>9}", v(i, j));
     }
     os << (i != v.rows() - 1 ? " |\n" : " |");
@@ -54,7 +54,7 @@ std::ostream& operator<<(std::ostream& os, const std::experimental::math::matrix
 }
 
 
-using namespace std::experimental::math;
+using namespace std::math;
 using namespace units::physical;
 using namespace units::physical::si::literals;